bubble sort

+++ 0-bubble_sort.c
@@ -1,6 +1,5 @@
 #include "sort.h"
 
-
 /**
  * swap - swap 2 element in an array
  * @ar: array
@@ -28,9 +27,9 @@ void bubble_sort(int *array, size_t size)
 {
 	int i, j;
 
-	for (i = 0; i < size - 1; i++)
+	for (i = 0; i < (int) size - 1; i++)
 	{
-		for (j = 0; j < size - 1; j++)
+		for (j = 0; j < (int) size - 1; j++)
 		{
 			if (array[j] > array[j + 1])
 			{
